+++ src/services/exercise.ts
@@ -6,12 +6,15 @@ export interface ExercisePostRequest {
   name: string
   description?: string
   muscle_groups?: string[]
+  property_ids: number[]
 }
 
 export interface ExercisePatchRequest extends PatchBase {
   name?: string
   description?: string
+  difficulty?: string
   muscle_groups?: string[]
+  property_ids?: number[]
 }
 
 export interface ExerciseQueryParams {