+++ src/types/category.ts
@@ -1,0 +1,10 @@
+import type { Entity } from "./base"
+import type { Property } from "../types/property"
+
+export interface Category extends Entity {
+  name: string
+}
+
+export interface CategoryProperties extends Category {
+  properties: Property[]
+}