+++ CHANGELOG.md
@@ -1,4 +1,20 @@
-# Version 0.2.0
+# Version 0.3.0 - 2025-07-27
+
+## Added
+
+- Admin menu to add, remove and change categories and its properties
+- Search filters for exercise categories and properties
+
+## Changed
+
+- Exercise detail page to inlcude admin defined categories & properties
+- Search help dialog to include exercise categories
+
+## Removed
+
+- Remove muscle groups from exercises in favor of new categories
+
+# Version 0.2.0 - 2025-07-20
 
 ## Added
 
@@ -8,7 +24,7 @@
 
 - Change carousel arrows position, scroll direction and style
 
-# Version 0.1.0
+# Version 0.1.0 - 2025-07-19
 
 ## Added